+++ docs/storybook/stories/forms/Select.stories.tsx
@@ -1,10 +1,13 @@
-import * as yup from 'yup';
 import { Button } from '@ttoss/ui';
-import { Form, FormFieldSelect } from '@ttoss/forms';
+import {
+  Form,
+  FormFieldSelect,
+  useForm,
+  yup,
+  yupResolver,
+} from '@ttoss/forms/src';
 import { Meta, Story } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
 
 export default {
   title: 'Forms/FormFieldSelect',
@@ -39,4 +42,29 @@ const Template: Story = () => {
   );
 };
 
+const Template2: Story = () => {
+  const formMethods = useForm({
+    mode: 'all',
+    resolver: yupResolver(schema),
+  });
+
+  return (
+    <Form {...formMethods} onSubmit={action('onSubmit')}>
+      <FormFieldSelect
+        name="car"
+        label="What car do you prefer?"
+        options={[
+          { value: 'Ferrari', label: 'Ferrari' },
+          { value: 'Mercedes', label: 'Mercedes' },
+          { value: 'BMW', label: 'BMW' },
+        ]}
+        placeholder="Please select a car"
+      />
+
+      <Button type="submit">Submit</Button>
+    </Form>
+  );
+};
+
 export const Example = Template.bind({});
+export const WithPlaceholder = Template2.bind({});