+++ packages/eslint-config/config.js
@@ -62,6 +62,17 @@ module.exports = {
     'formatjs/no-id': 'error',
     'formatjs/no-complex-selectors': 'error',
     'no-console': 'error',
+    /**
+     * https://github.com/jsx-eslint/eslint-plugin-react/issues/2628#issuecomment-984160944
+     */
+    'no-restricted-syntax': [
+      'error',
+      {
+        selector:
+          "ImportDeclaration[source.value='react'][specifiers.0.type='ImportDefaultSpecifier']",
+        message: 'Use "import * as React from \'react\'" instead.',
+      },
+    ],
     'no-use-before-define': ['error'],
     'prefer-arrow-callback': 'error',
     /**