+++ packages/forms/src/FormFieldSelect.tsx
@@ -1,10 +1,36 @@
-import { Box, Label, Select, type SelectProps } from '@ttoss/ui';
 import { ErrorMessage } from './ErrorMessage';
 import { FieldPath, FieldValues, useController } from 'react-hook-form';
+import { Flex, Label, Select, type SelectProps } from '@ttoss/ui';
 
 type FormRadioOption = {
   value: string | number;
   label: string;
+};
+
+type SelectSwitchProps =
+  | (SelectProps & { placeholder?: never })
+  | (SelectProps & { defaultValue?: never });
+
+const checkDefaultValue = (
+  options: Array<FormRadioOption>,
+  defaultValue?: string | number | readonly string[],
+  placeholder?: string
+) => {
+  const hasEmptyValue = options.some((opt) => {
+    return opt.value === '' || opt.value === 0;
+  });
+
+  if (placeholder && hasEmptyValue) return '';
+  if (placeholder && !hasEmptyValue) {
+    options.unshift({
+      label: placeholder,
+      value: '',
+    });
+    return '';
+  }
+  if (!placeholder && defaultValue) return defaultValue;
+  if (options.length === 0) return '';
+  return options?.[0]?.value;
 };
 
 export const FormFieldSelect = <
@@ -13,23 +39,32 @@ export const FormFieldSelect = <
   label,
   name,
   options,
+  sx,
   ...selectProps
 }: {
   label?: string;
   name: FieldPath<TFieldValues>;
   options: FormRadioOption[];
-} & SelectProps) => {
+} & SelectSwitchProps) => {
+  const { defaultValue, placeholder } = selectProps;
+
+  const checkedDefaultValue = checkDefaultValue(
+    options,
+    defaultValue,
+    placeholder
+  );
+
   const {
     field: { onChange, onBlur, value, ref },
   } = useController<any>({
     name,
-    defaultValue: '',
+    defaultValue: checkedDefaultValue,
   });
 
   const id = `form-field-select-${name}`;
 
   return (
-    <Box>
+    <Flex sx={{ flexDirection: 'column', width: '100%', ...sx }}>
       {label && <Label htmlFor={id}>{label}</Label>}
       <Select
         ref={ref}
@@ -38,7 +73,7 @@ export const FormFieldSelect = <
         onBlur={onBlur}
         value={value}
         id={id}
-        {...selectProps}
+        {...{ ...selectProps, defaultValue: undefined }}
       >
         {options.map((option) => {
           return (
@@ -49,6 +84,6 @@ export const FormFieldSelect = <
         })}
       </Select>
       <ErrorMessage name={name} />
-    </Box>
+    </Flex>
   );
 };