+++ packages/forms/tests/unit/FormFieldSelect.test.tsx
@@ -1,6 +1,7 @@
 import { Button } from '@ttoss/ui';
 import { Form, FormFieldSelect, useForm, yup, yupResolver } from '../../src';
 import { render, screen, userEvent } from '@ttoss/test-utils';
+import { useEffect, useState } from 'react';
 
 const RADIO_OPTIONS = [
   { value: '', label: 'Select a car' },
@@ -65,4 +66,175 @@ test('should display error messages', async () => {
   await user.click(screen.getByText('Submit'));
 
   expect(await screen.findByText('Car is required')).toBeInTheDocument();
+});
+
+test('should set a default value', async () => {
+  const user = userEvent.setup({ delay: null });
+
+  const onSubmit = jest.fn();
+
+  const RenderForm = () => {
+    const formMethods = useForm();
+
+    return (
+      <Form {...formMethods} onSubmit={onSubmit}>
+        <FormFieldSelect
+          name="car"
+          label="Cars"
+          options={RADIO_OPTIONS}
+          defaultValue="Ferrari"
+        />
+        <Button type="submit">Submit</Button>
+      </Form>
+    );
+  };
+
+  render(<RenderForm />);
+
+  await user.click(screen.getByText('Submit'));
+  expect(onSubmit).toHaveBeenCalledWith({ car: 'Ferrari' });
+});
+
+test('should have a default a value and change correctly', async () => {
+  const user = userEvent.setup({ delay: null });
+
+  const onSubmit = jest.fn();
+
+  const RenderForm = () => {
+    const formMethods = useForm();
+
+    return (
+      <Form {...formMethods} onSubmit={onSubmit}>
+        <FormFieldSelect
+          name="car"
+          label="Cars"
+          options={RADIO_OPTIONS}
+          defaultValue="Ferrari"
+        />
+        <Button type="submit">Submit</Button>
+      </Form>
+    );
+  };
+
+  render(<RenderForm />);
+
+  await user.selectOptions(
+    screen.getByRole('combobox'),
+    screen.getByText('BMW')
+  );
+
+  await user.click(screen.getByText('Submit'));
+
+  expect(onSubmit).toHaveBeenCalledWith({ car: 'BMW' });
+});
+
+test('should have an empty default when set a placeholder', async () => {
+  const RADIO_OPTIONS = [
+    { value: 'Ferrari', label: 'Ferrari' },
+    { value: 'Mercedes', label: 'Mercedes' },
+    { value: 'BMW', label: 'BMW' },
+  ];
+
+  const user = userEvent.setup({ delay: null });
+
+  const onSubmit = jest.fn();
+
+  const RenderForm = () => {
+    const formMethods = useForm();
+
+    return (
+      <Form {...formMethods} onSubmit={onSubmit}>
+        <FormFieldSelect
+          name="car"
+          label="Cars"
+          options={RADIO_OPTIONS}
+          placeholder="Select a car"
+        />
+        <Button type="submit">Submit</Button>
+      </Form>
+    );
+  };
+
+  render(<RenderForm />);
+
+  await user.click(screen.getByText('Submit'));
+
+  expect(
+    RADIO_OPTIONS.some((opt) => {
+      return opt.value === '';
+    })
+  ).toBeTruthy();
+
+  expect(onSubmit).toHaveBeenCalledWith({ car: '' });
+});
+
+test('should have the first option as default when nor placeholder, defaultValue or empty value is set', async () => {
+  const RADIO_OPTIONS = [
+    { value: 'Ferrari', label: 'Ferrari' },
+    { value: 'Mercedes', label: 'Mercedes' },
+    { value: 'BMW', label: 'BMW' },
+  ];
+
+  const user = userEvent.setup({ delay: null });
+
+  const onSubmit = jest.fn();
+
+  const RenderForm = () => {
+    const formMethods = useForm();
+
+    return (
+      <Form {...formMethods} onSubmit={onSubmit}>
+        <FormFieldSelect name="car" label="Cars" options={RADIO_OPTIONS} />
+        <Button type="submit">Submit</Button>
+      </Form>
+    );
+  };
+
+  render(<RenderForm />);
+
+  await user.click(screen.getByText('Submit'));
+
+  expect(onSubmit).toHaveBeenCalledWith({ car: 'Ferrari' });
+});
+
+test('When fetching, should display values correctly', async () => {
+  const RADIO_OPTIONS = [
+    { value: 'Ferrari', label: 'Ferrari' },
+    { value: 'Mercedes', label: 'Mercedes' },
+    { value: 'BMW', label: 'BMW' },
+  ];
+
+  const user = userEvent.setup({ delay: null });
+
+  const onSubmit = jest.fn();
+
+  const RenderForm = () => {
+    const formMethods = useForm();
+    const { resetField } = formMethods;
+    const [formOptions, setFormOptions] = useState<
+      {
+        value: string;
+        label: string;
+      }[]
+    >([]);
+
+    useEffect(() => {
+      setFormOptions(RADIO_OPTIONS);
+      // fetch are side effects, so, if the options depends on fetch and have a default value, the field should be reseted in the effect
+      resetField('car', { defaultValue: 'Ferrari' });
+    }, []);
+
+    return (
+      <Form {...formMethods} onSubmit={onSubmit}>
+        <FormFieldSelect name="car" label="Cars" options={formOptions} />
+        <Button type="submit">Submit</Button>
+      </Form>
+    );
+  };
+
+  render(<RenderForm />);
+
+  await user.click(screen.getByText('Submit'));
+
+  expect(onSubmit).toHaveBeenCalledWith({ car: 'Ferrari' });
 });