+++ packages/ui/src/components/CloseButton.tsx
@@ -1,27 +1,31 @@
+import * as React from 'react';
 import { Button, type ButtonProps } from 'theme-ui';
 import { Icon } from './Icon';
-import React from 'react';
 
 export type CloseButtonProps = ButtonProps & {
   label?: string;
+  onlyText?: boolean;
 };
 
 export const CloseButton = React.forwardRef<
   HTMLButtonElement,
   CloseButtonProps
->(({ label, sx, ...props }, ref) => {
+>(({ label, onlyText, ...props }, ref) => {
+  if (onlyText && !label) {
+    return null;
+  }
+
   return (
     <Button
       variant="buttons.closeButton"
       type="button"
       aria-label={label}
-      sx={{ gap: 'sm', padding: 0, ...sx }}
       {...props}
       ref={ref}
     >
       {label}
 
-      <Icon icon="close" />
+      {!onlyText && <Icon icon="close" />}
     </Button>
   );
 });