+++ packages/components/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.27.1](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.27.0...@ttoss/components@1.27.1) (2023-04-28)
+
+**Note:** Version bump only for package @ttoss/components
+
+# [1.27.0](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.44...@ttoss/components@1.27.0) (2023-04-28)
+
+### Features
+
+- add retain to critical resources ([#312](https://github.com/ttoss/ttoss/issues/312)) ([288e9de](https://github.com/ttoss/ttoss/commit/288e9de4021f7b8109487e593d5a55c8f4798b92))
+
+## [1.26.44](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.43...@ttoss/components@1.26.44) (2023-04-27)
+
+**Note:** Version bump only for package @ttoss/components
+
+## [1.26.43](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.42...@ttoss/components@1.26.43) (2023-04-27)
+
+**Note:** Version bump only for package @ttoss/components
+
+## [1.26.42](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.41...@ttoss/components@1.26.42) (2023-04-26)
+
+**Note:** Version bump only for package @ttoss/components
+
+## [1.26.41](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.40...@ttoss/components@1.26.41) (2023-04-25)
+
+**Note:** Version bump only for package @ttoss/components
+
+## [1.26.40](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.39...@ttoss/components@1.26.40) (2023-04-24)
+
+**Note:** Version bump only for package @ttoss/components
+
 ## [1.26.39](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.38...@ttoss/components@1.26.39) (2023-04-21)
 
 **Note:** Version bump only for package @ttoss/components