+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.27](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.26...@ttoss/theme@1.4.27) (2023-04-28)
+
+**Note:** Version bump only for package @ttoss/theme
+
+## [1.4.26](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.25...@ttoss/theme@1.4.26) (2023-04-28)
+
+**Note:** Version bump only for package @ttoss/theme
+
+## [1.4.25](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.24...@ttoss/theme@1.4.25) (2023-04-27)
+
+**Note:** Version bump only for package @ttoss/theme
+
+## 1.4.24 (2023-04-27)
+
+**Note:** Version bump only for package @ttoss/theme
+
+## [1.4.23](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.22...@ttoss/theme@1.4.23) (2023-04-24)
+
+**Note:** Version bump only for package @ttoss/theme
+
 ## [1.4.22](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.21...@ttoss/theme@1.4.22) (2023-04-21)
 
 **Note:** Version bump only for package @ttoss/theme