+++ terezinha-farm/auth/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.27](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.26...@terezinha-farm/auth@0.2.27) (2023-04-28)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## [0.2.26](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.25...@terezinha-farm/auth@0.2.26) (2023-04-26)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## 0.2.25 (2023-04-25)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
 ## [0.2.24](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.23...@terezinha-farm/auth@0.2.24) (2023-04-05)
 
 **Note:** Version bump only for package @terezinha-farm/auth